protect(unmap/map) pthread guard.
Due to mprotect() currently being a no-op the pthread guard page is not
protected. This adds calls to and an implementation of a function:

	void uk_mprotect_pthread_guard(addr, size, protect);

which can unmap/map the guard page when a thread is created/exits.
Unfortunately, the paging API does not seem to support reserving virtual
or physical memory. Therefore, we unmap/map this memory.

Issue: OS-231